+++ MultiRepoApp/RepositoryListModule/Model/RepositoryTableViewCellData.swift
@@ -1,0 +1,11 @@
+/// The data needed to configure a repository table view cell.
+struct RepositoryTableViewCellData {
+    /// The URL of the user's icon associated with the repository.
+    var userIcon: String
+    /// The name of the repository.
+    var title: String
+    /// The description of the repository.
+    var description: String
+    /// The source of the repository.
+    var source: String
+}