+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `RepositoryTableViewCell` implementation.
 - `SearchTableViewCell` implementation.
 - `RepositorySearchDelegateProtocol` implementation.
+- `RepositoryListViewController` implementation.
+- `RepositoryListTableViewSection` implementation.
 
 ## [0.1.0] - 2023-05-02